@@ -16,7 +16,7 @@ class TemplateFactory extends Factory
     public function definition(): array
     {
         return [
-            'name' => $this->faker->word,
+            'name' => $this->faker->word(),
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'content' => '{{content}}'
         ];
