@@ -192,13 +192,13 @@ class CampaignControllerTest extends TestCase
         $campaign = Campaign::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $campaignUpdateData = [
-            'name' => $this->faker->word,
-            'subject' => $this->faker->sentence,
-            'from_name' => $this->faker->name,
-            'from_email' => $this->faker->safeEmail,
+            'name' => $this->faker->word(),
+            'subject' => $this->faker->sentence(),
+            'from_name' => $this->faker->name(),
+            'from_email' => $this->faker->safeEmail(),
             'email_service_id' => $campaign->email_service_id,
             'template_id' => $campaign->template_id,
-            'content' => $this->faker->paragraph
+            'content' => $this->faker->paragraph()
         ];
 
         // when
@@ -318,13 +318,13 @@ class CampaignControllerTest extends TestCase
         $template = Template::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         return array_merge([
-            'name' => $this->faker->word,
-            'subject' => $this->faker->sentence,
-            'from_name' => $this->faker->name,
-            'from_email' => $this->faker->safeEmail,
+            'name' => $this->faker->word(),
+            'subject' => $this->faker->sentence(),
+            'from_name' => $this->faker->name(),
+            'from_email' => $this->faker->safeEmail(),
             'email_service_id' => $emailService->id,
             'template_id' => $template->id,
-            'content' => $this->faker->paragraph
+            'content' => $this->faker->paragraph()
         ], $overrides);
     }
 }
